web development
Hey Everyone!!
It's my first front-end webpage, which I built up using HTML, CSS, and some parts of JS. This content has all the resources that I follow further, and it is more helpful to those who start their journey in the field of web development. It content everything which really need a web developer.